FINAL CRUD. registrar, visualizar, eliminar y actualizar

+++ gerente/views.py
@@ -1,11 +1,11 @@
+import os
 from django.shortcuts import render, redirect
 from .forms import LoginForm, EmpleadoForm
 from .models import empleado
 from django.http import JsonResponse
 from .models import empleado  # Asegúrate de importar tu modelo empleado correctamente
 import base64
-from django.conf import settings
-import os
+
 
 def iniciar_sesion(request):
     if request.method == 'POST':
@@ -48,6 +48,7 @@ def crear_empleado(request):
 
 def consultar_admin(request):
     return render(request, 'consultar_admin.html')
+
 
 
 def buscar_empleado(request):
@@ -116,20 +117,23 @@ def editar_empleado(request):
     
 
 
+
 def eliminar_empleado(request):
     if request.method == 'GET':
         cedula = request.GET.get('cedula')  # Obtener la cédula del empleado a eliminar
         try:
             # Buscar el empleado por su cédula
             empleado_obj = empleado.objects.get(cedula=cedula)
-            # Obtener la ruta del archivo de imagen
-            ruta_imagen = empleado_obj.foto_emp.path
+            
+            # Eliminar el archivo de imagen si existe
+            if empleado_obj.foto_emp:
+                ruta_imagen = empleado_obj.foto_emp.path
+                if os.path.exists(ruta_imagen):
+                    os.remove(ruta_imagen)
+            
             # Eliminar el objeto empleado de la base de datos
             empleado_obj.delete()
-            # Eliminar el archivo de imagen de la carpeta de medios si existe
-            if os.path.exists(ruta_imagen):
-                os.remove(ruta_imagen)
-                           
+            
             return JsonResponse({'success': True, 'message': 'Empleado eliminado correctamente'})
         except empleado.DoesNotExist:
             return JsonResponse({'success': False, 'message': 'El empleado no existe'})
@@ -137,5 +141,3 @@ def eliminar_empleado(request):
             return JsonResponse({'success': False, 'message': str(e)})
     else:
         return JsonResponse({'error': 'Método no permitido'}, status=405)
-
-
